+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2019-04-24
+### Added
+
 ## [0.3.0] - 2019-03-14
 ### Added
 - Added initial iOS platform support (#1)
@@ -32,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `LuaThread#execAsset()` method
 - `LuaThread#execFile()` method
 
+[0.3.1]:  https://github.com/drydart/flutter_lua/compare/0.3.0...0.3.1
 [0.3.0]:  https://github.com/drydart/flutter_lua/compare/0.2.0...0.3.0
 [0.2.0]:  https://github.com/drydart/flutter_lua/compare/0.1.0...0.2.0
 [0.1.0]:  https://github.com/drydart/flutter_lua/compare/0.0.0...0.1.0